+++ tools/mo/openvino/tools/mo/convert_impl.py
@@ -11,8 +11,6 @@ from collections import OrderedDict
 from copy import deepcopy
 from pathlib import Path
 
-import numpy as np
-
 try:
     import openvino_telemetry as tm
 except ImportError:
@@ -32,8 +30,8 @@ from openvino.tools.mo.utils.cli_parser import check_available_transforms, \
     get_advanced_cli_options, get_available_front_ends, get_caffe_cli_options, \
     get_common_cli_options, get_freeze_placeholder_values, get_kaldi_cli_options, get_layout_values, \
     get_mean_scale_dictionary, get_mxnet_cli_options, get_onnx_cli_options, \
-    get_placeholder_shapes, get_tf_cli_options, get_tuple_values, parse_transform, parse_tuple_pairs, \
-    get_all_cli_parser, mo_convert_params, get_model_name_from_args, split_shapes, depersonalize
+    get_placeholder_shapes, get_tf_cli_options, parse_transform, parse_tuple_pairs, \
+    mo_convert_params, get_model_name_from_args, depersonalize
 
 from openvino.tools.mo.utils.error import Error
 from openvino.tools.mo.utils.find_ie_version import find_ie_version
@@ -44,12 +42,12 @@ from openvino.tools.mo.utils.telemetry_utils import send_params_info, send_frame
 from openvino.tools.mo.utils.version import get_simplified_mo_version, get_simplified_ie_version, get_version
 from openvino.tools.mo.utils.versions_checker import check_requirements  # pylint: disable=no-name-in-module
 from openvino.tools.mo.utils.telemetry_utils import get_tid
-from openvino.tools.mo.front.common.partial_infer.utils import mo_array
 from openvino.tools.mo.moc_frontend.check_config import legacy_extensions_used
+from openvino.tools.mo.moc_frontend.pytorch_frontend_utils import get_pytorch_decoder, convert_pytorch_via_onnx
+from openvino.tools.mo.moc_frontend.shape_utils import parse_input_shapes, get_static_shape
 
 # pylint: disable=no-name-in-module,import-error
 from openvino.frontend import FrontEndManager, OpConversionFailure, ProgressReporterExtension, TelemetryExtension
-from openvino.runtime import PartialShape, Dimension
 from openvino.runtime import get_version as get_rt_version
 
 
@@ -537,64 +535,6 @@ def emit_ir(graph: Graph, argv: argparse.Namespace, non_default_params: dict):
     return func
 
 
-def get_static_shape(shape: [PartialShape, list, tuple], dynamic_value=None):
-    # Current function returns list with static dimensions with following logic.
-    # For dynamic dimensions return lower boundaries if they are set, otherwise
-    # return upper boundaries if they are set. If dimension is fully dynamic then raise error.
-    shape_list = []
-    for idx, dim in enumerate(shape):
-        if isinstance(dim, int):
-            if dim == -1:
-                shape_list.append(dynamic_value)
-                continue
-            shape_list.append(dim)
-        elif isinstance(dim, np.int64):
-            if dim == np.int64(-1):
-                shape_list.append(dynamic_value)
-                continue
-            shape_list.append(dim)
-        elif isinstance(dim, tuple):
-            # tuple where (min_length, max_length), the format which uses MO cli parser
-            assert len(dim) == 2, "Unknown dimension type {}".format(dim)
-            if dim[0] > 0:
-                shape_list.append(dim[0])
-            elif dim[1] < np.iinfo(np.int64).max:
-                shape_list.append(dim[1])
-            else:
-                shape_list.append(dynamic_value)
-                continue
-        elif isinstance(dim, Dimension):
-            if dim.is_static or dim.get_min_length() > 0:
-                shape_list.append(dim.get_min_length())
-            elif dim.get_max_length() != -1:
-                shape_list.append(dim.get_max_length())
-            else:
-                shape_list.append(dynamic_value)
-                continue
-        else:
-            raise Error("Unknown dimension type {}".format(dim))
-
-    return tuple(shape_list)
-
-
-def get_dynamic_dims(shape: [PartialShape, list, tuple]):
-    dynamic_dims = []
-    for idx, dim in enumerate(shape):
-        if isinstance(dim, int):
-            if dim == -1:
-                dynamic_dims.append(idx)
-        if isinstance(dim, np.int64):
-            if dim == np.int64(-1):
-                dynamic_dims.append(idx)
-        elif isinstance(dim, tuple):
-            dynamic_dims.append(idx)
-        elif isinstance(dim, Dimension):
-            if dim.get_min_length() == 0 and dim.get_max_length() == -1:
-                dynamic_dims.append(idx)
-
-    return dynamic_dims
-
-
 def check_model_object(argv):
     model = argv['input_model']
     if 'tensorflow' in sys.modules:
@@ -646,163 +586,6 @@ def check_model_object(argv):
         return 'onnx'
 
     raise Error('Unknown model type: {}'.format(type(model)))
-
-
-def get_onnx_temp_filename(output_dir):
-    output_dir = output_dir if output_dir is not None else os.getcwd()
-    return os.path.normpath(os.path.join(output_dir, "model.onnx"))
-
-
-def to_torch_tensor(tensor):
-    import torch
-    from openvino.runtime import Tensor
-    if isinstance(tensor, torch.Tensor):
-        return tensor
-    if isinstance(tensor, np.ndarray):
-        return torch.tensor(tensor)
-    if isinstance(tensor, np.ndarray):
-        return torch.tensor(tensor)
-    if isinstance(tensor, Tensor):
-        return torch.tensor(tensor.data)
-    else:
-        raise Error("Unexpected type of example_input. Supported types torch.Tensor, np.array or ov.Tensor. "
-                    "Got {}".format(type(tensor)))
-
-
-def prepare_torch_inputs(example_inputs, input_shape, allow_none=False):
-    import torch
-    inputs = None
-    if example_inputs is not None:
-        inputs = example_inputs
-        if isinstance(inputs, list):
-            inputs = [to_torch_tensor(x) for x in inputs]
-            if len(inputs) == 1:
-                inputs = torch.unsqueeze(inputs[0], 0)
-            else:
-                inputs = inputs
-        elif isinstance(inputs, tuple):
-            inputs = [to_torch_tensor(x) for x in inputs]
-            inputs = tuple(inputs)
-        elif isinstance(inputs, dict):
-            for name, tensor in inputs.items():
-                assert isinstance(name, str), "Expected dictionary where keys are input names of string type and" \
-                                              " values are tensors. Got key of type {}".format(type(name))
-                inputs[name] = to_torch_tensor(tensor)
-        else:
-            inputs = to_torch_tensor(inputs)
-    elif input_shape is not None:
-        inputs = []
-        for shape in input_shape:
-            static_shape = get_static_shape(shape, dynamic_value=1)
-            inputs.append(torch.zeros(static_shape))
-        inputs = tuple(inputs)
-    else:
-        if not allow_none:
-            raise Error("Please provide input_shape or example_input for converting PyTorch model.")
-    return inputs
-
-
-def convert_pytorch_to_onnx(model, input_shape, opset_version, example_inputs, output_dir):
-    import io
-    import torch
-
-    input_names = None
-    inputs = prepare_torch_inputs(example_inputs, input_shape)
-
-    dynamic_dims_dict = {}
-    if input_shape is not None and input_names is None:
-        input_names = ["input_{}".format(idx) for idx in range(len(input_shape))]
-        for shape_idx, shape in enumerate(input_shape):
-            dynamic_dims = get_dynamic_dims(shape)
-            if len(dynamic_dims) > 0:
-                dynamic_dims_dict[input_names[shape_idx]] = dynamic_dims
-    additional_params = {}
-    if len(dynamic_dims_dict) > 0:
-        additional_params.update({'dynamic_axes': dynamic_dims_dict})
-    if input_names is not None and len(input_names) > 0:
-        additional_params.update({'input_names': input_names})
-
-    if os.environ.get('SAVE_TO_BYTES_IO_ONNX_MODEL'):
-        model_onnx = io.BytesIO()
-    else:
-        model_onnx = get_onnx_temp_filename(output_dir)
-    if opset_version is not None:
-        additional_params.update({'opset_version': opset_version})
-
-    torch.onnx.export(model,
-                      inputs,
-                      model_onnx,
-                      **additional_params)
-    return model_onnx
-
-
-def get_pytorch_decoder(model, input_shape, example_inputs):
-    import torch
-    import inspect
-    try:
-        from openvino.frontend.pytorch.decoder import TorchScriptPythonDecoder
-    except Exception as e:
-        print("PyTorch frontend loading failed")
-        raise e
-    inputs = prepare_torch_inputs(example_inputs, input_shape, allow_none=True)
-    model.eval()
-    input_signature = None
-    if isinstance(model, torch.nn.Module) and not isinstance(model, torch.jit._trace.TopLevelTracedModule):
-        input_signature = list(inspect.signature(model.forward).parameters.keys())
-        try:
-            scripted = torch.jit.script(model)
-        except Exception as scripting_err:
-            if example_inputs is not None:
-                try:
-                    scripted = torch.jit.trace(model, inputs)
-                except Exception as tracing_e:
-                    print('Both traicing and scripting failed')
-                    raise tracing_e
-            else:
-                print("Model scripting failed")
-                raise scripting_err
-    else:
-        scripted = model
-    f_model = torch.jit.freeze(scripted)
-    decoder = TorchScriptPythonDecoder(f_model)
-    return decoder, input_signature
-
-
-def parse_input_shapes(argv):
-    input_shapes = None
-    if 'input_shape' in argv and argv['input_shape'] is not None:
-        shapes = argv['input_shape']
-        if isinstance(shapes, str):
-            shapes = ["[{}]".format(x) for x in split_shapes(shapes)]
-        if isinstance(shapes, list) or isinstance(shapes, tuple):
-            input_shapes = []
-            is_single_shape = False
-            for shape in shapes:
-                if isinstance(shape, str):
-                    _, shape_tuple, _ = get_placeholder_shapes(argv_input=None, argv_input_shape=shape)
-                    input_shapes.append(shape_tuple)
-                    if is_single_shape:
-                        raise Error("Incorrect format of shape.")
-                elif isinstance(shape, int) or isinstance(shape, np.int64) or isinstance(shape, Dimension):
-                    is_single_shape = True
-                    input_shapes.append(shape)
-                else:
-                    input_shapes.append(shape)
-            if is_single_shape:
-                return [input_shapes]
-            else:
-                return input_shapes
-        elif isinstance(shapes, PartialShape):
-            return [shapes]
-        else:
-            try:
-                import torch
-                if isinstance(shapes, torch.Size):
-                    return [shapes]
-            except ImportError:
-                raise Error("Unknown type of input shape {}.".format(type(shapes)))
-
-    return input_shapes
 
 
 def driver(argv: argparse.Namespace, non_default_params: dict):
@@ -957,43 +740,6 @@ def pack_params_to_args_namespace(args: dict, cli_parser: argparse.ArgumentParse
     return argv
 
 
-def remove_tmp_onnx_model(out_dir):
-    if not os.environ.get('SAVE_TO_BYTES_IO_ONNX_MODEL'):
-        tmp_onnx_model = get_onnx_temp_filename(out_dir)
-
-        if os.path.exists(tmp_onnx_model):
-            os.remove(tmp_onnx_model)
-
-
-def convert_pytorch_via_onnx(args, example_inputs, cli_parser, framework):
-    opset_version = None
-    if 'onnx_opset_version' in args and args['onnx_opset_version'] is not None:
-        opset_version = args['onnx_opset_version']
-
-    out_dir = args['output_dir'] if 'output_dir' in args else None
-
-    model_onnx = convert_pytorch_to_onnx(args['input_model'],
-                                         parse_input_shapes(args),
-                                         opset_version,
-                                         example_inputs,
-                                         out_dir)
-
-    args['input_model'] = model_onnx
-    if os.environ.get('SAVE_TO_BYTES_IO_ONNX_MODEL'):
-        args['use_legacy_frontend'] = True
-    args['example_input'] = None
-    args['onnx_opset_version'] = None
-
-    try:
-        ov_model, argv = _convert(cli_parser, framework, args)
-    except Exception as e:
-        remove_tmp_onnx_model(out_dir)
-        raise e
-
-    remove_tmp_onnx_model(out_dir)
-    return ov_model, argv
-
-
 def _convert(cli_parser: argparse.ArgumentParser, framework, args):
     if 'help' in args and args['help']:
         show_mo_convert_help()
@@ -1002,30 +748,29 @@ def _convert(cli_parser: argparse.ArgumentParser, framework, args):
     telemetry = tm.Telemetry(tid=get_tid(), app_name='Model Optimizer', app_version=get_simplified_mo_version())
     telemetry.start_session('mo')
     telemetry.send_event('mo', 'version', get_simplified_mo_version())
+    # Initialize logger with 'ERROR' as default level to be able to form nice messages
+    # before arg parser deliver log_level requested by user
+    init_logger('ERROR', False)
     try:
         model_framework = None
         inp_model_is_object = input_model_is_object(args)
         if inp_model_is_object:
             model_framework = check_model_object(args)
             if model_framework == "pytorch":
-                use_pt_fe = not args.get("use_legacy_frontend", False)
                 example_inputs = None
                 if 'example_input' in args and args['example_input'] is not None:
                     example_inputs = args['example_input']
-                if not use_pt_fe:
+                   
+                if 'use_legacy_frontend' in args and args['use_legacy_frontend']:
                     # TO DO: remove this path, when pytorch frontend productization is finished, CVS-103726
-                    if "use_legacy_frontend" in args:
-                        args.pop("use_legacy_frontend")
-                    return convert_pytorch_via_onnx(args, example_inputs, cli_parser, framework)
+                    # prevent invoking legacy mo python onnx frontend for models converted on the fly
+                    args.pop("use_legacy_frontend")
+                    return convert_pytorch_via_onnx(args, example_inputs, cli_parser, framework, _convert)
 
                 decoder, input_signature  = get_pytorch_decoder(args['input_model'], parse_input_shapes(args), example_inputs)
                 args['input_model'] = decoder
                 args["framework"] = "pytorch"
                 args["input_signature"] = input_signature
-
-        # Initialize logger with 'ERROR' as default level to be able to form nice messages
-        # before arg parser deliver log_level requested by user
-        init_logger('ERROR', False)
 
         argv = pack_params_to_args_namespace(args, cli_parser)
